fix: tets-get_star
